Design issues: Menu:
Seeking Alpha marketplace logo is too big.
Should be 245x28pt with 16pt left alignment and 28pt align from top.
All list including profile pictures should be 16pt aligned from left
“Logout” should be 16pt aligned from right
Notification badge should be round and not diamond.
Should be 15pt diameter and font inside is 10pt.
should be 16pt aligned from right.
Headlines (unreads, subscription, etc.) should be inside a 30pt height cell not including the dividers. Font is 12pt all caps.
The dividers in headline cells, between the darker background of the cells, should all be 1pt.
List cells should all be 48pt height.
Text in cells should be centered from top to bottom of cells.

Name of the service on the nav bar should be in bold font and centered aligned to the app (right now it’s aligned to left so it’s not centered and too close to the right side of the navbar). Also the name of service is cropped at the bottom, notice the “g”.
Profile pictures should be 30pt size and 16pt aligned from left.